Move support functions to core
