+++ packages/core/src/util/schema.ts
@@ -48,3 +48,11 @@ export const isOneOfEnumSchema = (schema: JsonSchema) =>
   Object.prototype.hasOwnProperty.call(schema, 'oneOf') &&
   schema.oneOf &&
   (schema.oneOf as JsonSchema[]).every((s) => s.const !== undefined);
+
+/**
+ * Tests whether the schema has an enum.
+ */
+export const isEnumSchema = (schema: JsonSchema) =>
+  !!schema &&
+  ((Object.prototype.hasOwnProperty.call(schema, 'enum') && schema.enum) ||
+    (Object.prototype.hasOwnProperty.call(schema, 'const') && schema.const));
